TaskRunner, an abstraction over ExecutorService
I want to tighten up our executors for a few reasons
 - Fix daemon vs. non-daemon problems
 - Fix code unloading problems
 - Be able to wait for async work to complete in tearDown()
 - Fewer threads for pongs, acks, and window updates

This shouldn't be too much to layer on top of ExecutorService, but
it is not easy to do with those APIs alone.

The underlying motivation is to make scheduling work easier in
Http2Connection, which will help with differentiating between
dead streams and dead connections.

https://github.com/square/okhttp/issues/3146